Include note from postgresql upgrading notes
Note that I simplified the `kubectl logs` command a lot. This is
slightly less general than choosing the name based off of the pods using
the label selector like they do in the original notes, but it goes more
in line with the rest of the commands in this document. And some
modification was needed anyway, becuase the instance label didn't apply
in our case, so I preferred to make it simpler.

Signed-off-by: Andreas Lindhé <andreas@lindhe.io>